Use Hex encoding in RPC methods (#752)
* Clean RPC types, use ethereum encodings

* Fix failing test

+++ tests/bundles/bundle_sponsored_test.go
@@ -111,8 +111,8 @@ func TestBundle_CanRunSponsorshipAndSponsored(t *testing.T) {
 	// 2. the sponsored transaction
 	// 3. the internal transaction that transfer the fee from the sponsee to the sponsor
 	txs := block.Transactions()
-	position := *info.Position
-	require.GreaterOrEqual(t, uint32(len(txs)), position+3)
+	position := uint(*info.Position)
+	require.GreaterOrEqual(t, uint(len(txs)), position+3)
 	require.Equal(t, txs[position].Hash(), bundle.Transactions[0].Hash())
 	require.Equal(t, txs[position+1].Hash(), bundle.Transactions[1].Hash())
 	require.True(t, internaltx.IsInternal(txs[position+2]))
